time_get returns the correct value as a float.

+++ include/timer.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-	extern long time_get();
+	extern float time_get();
 
 #ifdef __cplusplus
 }
